修改：表信息相关接口和前端页面初步完成，后续有待需求文档与原型，再进行修改

+++ src/main/java/com/dili/ia/domain/dto/MeterDto.java
@@ -45,7 +45,7 @@ public class MeterDto extends Meter {
     /**
      * 表类型
      */
-    private Long type;
+    private Integer type;
 
     /**
      * 部门
@@ -68,7 +68,7 @@ public class MeterDto extends Meter {
     private Long assetsId;
 
     /**
-     * 对应编号,名称
+     * 对应编号,名称(表地址)
      */
     private String assetsName;
